+++ _travis/upload_coverage.sh
@@ -3,6 +3,6 @@
 set -exo pipefail
 
 if [[ -e .coverage ]]; then
-    python3.6 -m pip install codecov
-    python3.6 -m codecov --env TRAVIS_OS_NAME,NOX_SESSION
+    python3 -m pip install codecov
+    python3 -m codecov --env TRAVIS_OS_NAME,NOX_SESSION
 fi