+++ test/test_response.py
@@ -1,10 +1,13 @@
+# -*- coding: utf-8 -*-
+
 import socket
 import zlib
 
-from io import BytesIO, BufferedReader
+from io import BytesIO, BufferedReader, TextIOWrapper
 
 import pytest
 import mock
+import six
 
 from urllib3.response import HTTPResponse, brotli
 from urllib3.exceptions import (
@@ -349,6 +352,14 @@ class TestResponse(object):
         br.close()
         assert resp.closed
 
+        # HTTPResponse.read() by default closes the response
+        # https://github.com/urllib3/urllib3/issues/1305
+        fp = BytesIO(b"hello\nworld")
+        resp = HTTPResponse(fp, preload_content=False)
+        with pytest.raises(ValueError) as ctx:
+            list(BufferedReader(resp))
+        assert str(ctx.value) == "readline of closed file"
+
         b = b"fooandahalf"
         fp = BytesIO(b)
         resp = HTTPResponse(fp, preload_content=False)
@@ -361,6 +372,72 @@ class TestResponse(object):
         # gets tested.
         while not br.closed:
             br.read(5)
+
+    def test_io_not_autoclose_bufferedreader(self):
+        fp = BytesIO(b"hello\nworld")
+        resp = HTTPResponse(fp, preload_content=False, auto_close=False)
+        reader = BufferedReader(resp)
+        assert list(reader) == [b"hello\n", b"world"]
+
+        assert not reader.closed
+        assert not resp.closed
+        with pytest.raises(StopIteration):
+            next(reader)
+
+        reader.close()
+        assert reader.closed
+        assert resp.closed
+        with pytest.raises(ValueError) as ctx:
+            next(reader)
+        assert str(ctx.value) == "readline of closed file"
+
+    def test_io_textiowrapper(self):
+        fp = BytesIO(b"\xc3\xa4\xc3\xb6\xc3\xbc\xc3\x9f")
+        resp = HTTPResponse(fp, preload_content=False)
+        br = TextIOWrapper(resp, encoding="utf8")
+
+        assert br.read() == u"äöüß"
+
+        br.close()
+        assert resp.closed
+
+        # HTTPResponse.read() by default closes the response
+        # https://github.com/urllib3/urllib3/issues/1305
+        fp = BytesIO(
+            b"\xc3\xa4\xc3\xb6\xc3\xbc\xc3\x9f\n\xce\xb1\xce\xb2\xce\xb3\xce\xb4"
+        )
+        resp = HTTPResponse(fp, preload_content=False)
+        with pytest.raises(ValueError) as ctx:
+            if six.PY2:
+                # py2's implementation of TextIOWrapper requires `read1`
+                # method which is provided by `BufferedReader` wrapper
+                resp = BufferedReader(resp)
+            list(TextIOWrapper(resp))
+        assert str(ctx.value) == "I/O operation on closed file."
+
+    def test_io_not_autoclose_textiowrapper(self):
+        fp = BytesIO(
+            b"\xc3\xa4\xc3\xb6\xc3\xbc\xc3\x9f\n\xce\xb1\xce\xb2\xce\xb3\xce\xb4"
+        )
+        resp = HTTPResponse(fp, preload_content=False, auto_close=False)
+        if six.PY2:
+            # py2's implementation of TextIOWrapper requires `read1`
+            # method which is provided by `BufferedReader` wrapper
+            resp = BufferedReader(resp)
+        reader = TextIOWrapper(resp, encoding="utf8")
+        assert list(reader) == [u"äöüß\n", u"αβγδ"]
+
+        assert not reader.closed
+        assert not resp.closed
+        with pytest.raises(StopIteration):
+            next(reader)
+
+        reader.close()
+        assert reader.closed
+        assert resp.closed
+        with pytest.raises(ValueError) as ctx:
+            next(reader)
+        assert str(ctx.value) == "I/O operation on closed file."
 
     def test_streaming(self):
         fp = BytesIO(b"foo")
@@ -434,7 +511,7 @@ class TestResponse(object):
 
         class MockCompressedDataReading(BytesIO):
             """
-            A ByteIO-like reader returning ``payload`` in ``NUMBER_OF_READS``
+            A BytesIO-like reader returning ``payload`` in ``NUMBER_OF_READS``
             calls to ``read``.
             """