+++ CHANGES.rst
@@ -9,6 +9,14 @@ dev (master)
 * Fix edge case where Retry-After header was still respected even when
   explicitly opted out of. (Pull #1607)
 
+* Remove dependency on ``rfc3986`` for URL parsing.
+
+* Fix issue where URLs containing invalid characters within ``Url.auth`` would
+  raise an exception instead of percent-encoding those characters.
+
+* Add support for ``HTTPResponse.auto_close = False`` which makes HTTP responses
+  work well with BufferedReaders and other ``io`` module features. (Pull #1652)
+
 
 1.25.3 (2019-05-23)
 -------------------